+++ zntrack/fields/zn/__init__.py
@@ -11,14 +11,8 @@ import znflow.utils
 import zninit
 import znjson
 
-from zntrack.fields.field import Field
-from zntrack.utils import (
-    LazyOption,
-    NodeStatusResults,
-    config,
-    module_handler,
-    update_key_val,
-)
+from zntrack.fields.field import Field, LazyField
+from zntrack.utils import module_handler, update_key_val
 
 if typing.TYPE_CHECKING:
     from zntrack import Node
@@ -110,21 +104,11 @@ class Params(Field):
         with open(file, "w") as f:
             yaml.safe_dump(params_dict, f, indent=4)
 
-    def load(self, instance: "Node", lazy: bool = None):
-        """Load the field from disk.
-
-        Parameters
-        ----------
-        instance : Node
-            The node instance associated with this field.
-        lazy : bool, optional
-            Lazy does not apply to Params.
-        """
+    def _get_value_from_file(self, instance: "Node") -> any:
         file = self.get_affected_files(instance)[0]
         params_dict = yaml.safe_load(instance.state.get_file_system().read_text(file))
         value = params_dict[instance.name].get(self.name, None)
-        value = json.loads(json.dumps(value))
-        instance.__dict__[self.name] = value
+        return json.loads(json.dumps(value))
 
     def get_stage_add_argument(self, instance: "Node") -> typing.List[tuple]:
         """Get the DVC stage add argument for this field.
@@ -143,7 +127,7 @@ class Params(Field):
         return [(f"--{self.dvc_option}", f"{file}:{instance.name}")]
 
 
-class Output(Field):
+class Output(LazyField):
     """A field that is saved to disk."""
 
     def __init__(self, dvc_option: str, **kwargs):
@@ -191,37 +175,12 @@ class Output(Field):
             json.dumps(getattr(instance, self.name), cls=znjson.ZnEncoder, indent=4)
         )
 
-    def __get__(self, instance, owner=None):
-        """Getter method with added LazyOption support."""
-        if instance is None:
-            return self
-        if instance.__dict__[self.name] is LazyOption:
-            self.load(instance, lazy=False)
-
-        return super().__get__(instance, owner)
-
-    def load(self, instance: "Node", lazy: bool = None):
-        """Load the field from disk.
-
-        Parameters
-        ----------
-        instance : Node
-            The node instance.
-        lazy : bool, optional
-            If True, the field will be loaded lazily when it is accessed.
-        """
-        if lazy is None and config.lazy:
-            instance.__dict__[self.name] = LazyOption
-        else:
-            file = self.get_affected_files(instance)[0]
-            try:
-                value = json.loads(
-                    instance.state.get_file_system().read_text(file.as_posix()),
-                    cls=znjson.ZnDecoder,
-                )
-                instance.__dict__[self.name] = value
-            except FileNotFoundError:
-                instance.state.results = NodeStatusResults.UNKNOWN
+    def _get_value_from_file(self, instance: "Node") -> any:
+        file = self.get_affected_files(instance)[0]
+        return json.loads(
+            instance.state.get_file_system().read_text(file.as_posix()),
+            cls=znjson.ZnDecoder,
+        )
 
     def get_stage_add_argument(self, instance) -> typing.List[tuple]:
         """Get the DVC command for this field.
@@ -240,7 +199,7 @@ class Output(Field):
         return [(f"--{self.dvc_option}", file.as_posix())]
 
 
-class Plots(Field):
+class Plots(LazyField):
     """A field that is saved to disk."""
 
     dvc_option: str = "plots"
@@ -260,28 +219,11 @@ class Plots(Field):
         value: pd.DataFrame = getattr(instance, self.name)
         value.to_csv(file)
 
-    def __get__(self, instance, owner=None):
-        # TODO make this a mixin / parent method
-        if instance is None:
-            return self
-        if instance.__dict__[self.name] is LazyOption:
-            self.load(instance, lazy=False)
-
-        return super().__get__(instance, owner)
-
-    def load(self, instance: "Node", lazy: bool = None):
-        """Load the field from disk."""
-        if lazy is None and config.lazy:
-            instance.__dict__[self.name] = LazyOption
-        else:
-            file = self.get_affected_files(instance)[0]
-            try:
-                value = pd.read_csv(
-                    instance.state.get_file_system().open(file.as_posix()), index_col=0
-                )
-                instance.__dict__[self.name] = value
-            except FileNotFoundError:
-                instance.state.results = NodeStatusResults.UNKNOWN
+    def _get_value_from_file(self, instance: "Node") -> any:
+        file = self.get_affected_files(instance)[0]
+        return pd.read_csv(
+            instance.state.get_file_system().open(file.as_posix()), index_col=0
+        )
 
     def get_stage_add_argument(self, instance) -> typing.List[tuple]:
         """Get the dvc command for this field."""
@@ -292,7 +234,7 @@ class Plots(Field):
 _default = object()
 
 
-class Dependency(Field):
+class Dependency(LazyField):
     """A dependency field."""
 
     dvc_option = "deps"
@@ -349,38 +291,23 @@ class Dependency(Field):
             ),
         )
 
-    def __get__(self, instance, owner=None):
-        # TODO make this a mixin / parent method
-        if instance is None:
-            return self
-        if instance.__dict__[self.name] is LazyOption:
-            self.load(instance, lazy=False)
+    def _get_value_from_file(self, instance: "Node") -> any:
+        zntrack_dict = json.loads(
+            instance.state.get_file_system().read_text("zntrack.json"),
+        )
+        value = zntrack_dict[instance.name][self.name]
 
-        return super().__get__(instance, owner)
+        value = update_key_val(value, instance=instance)
 
-    def load(self, instance: "Node", lazy: bool = None):
-        """Load the field from disk."""
-        if lazy is None and config.lazy:
-            instance.__dict__[self.name] = LazyOption
-        else:
-            zntrack_dict = json.loads(
-                instance.state.get_file_system().read_text("zntrack.json"),
-            )
-            value = zntrack_dict[instance.name][self.name]
+        value = json.loads(
+            json.dumps(value),
+            cls=znjson.ZnDecoder.from_converters(ConnectionConverter, add_default=True),
+        )
 
-            value = update_key_val(value, instance=instance)
-
-            value = json.loads(
-                json.dumps(value),
-                cls=znjson.ZnDecoder.from_converters(
-                    ConnectionConverter, add_default=True
-                ),
-            )
-
-            # Up until here we have connection objects. Now we need
-            # to resolve them to Nodes. The Nodes, as in 'connection.instance'
-            #  are already loaded by the ZnDecoder.
-            instance.__dict__[self.name] = znflow.graph._UpdateConnectors()(value)
+        # Up until here we have connection objects. Now we need
+        # to resolve them to Nodes. The Nodes, as in 'connection.instance'
+        #  are already loaded by the ZnDecoder.
+        return znflow.graph._UpdateConnectors()(value)
 
     def get_stage_add_argument(self, instance) -> typing.List[tuple]:
         """Get the dvc command for this field."""
